Rough first pass at removing parsing logic from lexer
The original lexer was doing almost all the parsing leaving the parser
to do almost nothing. While this was 'working', it would have made
things more difficult in the future as more advanced features were
added.

It also would prevent us from doing any kind of error recovery during
parsing as that needs to be done in the parser and not the lexer.